+++ databento/live/session.py
@@ -8,7 +8,6 @@ import struct
 import threading
 from collections.abc import Iterable
 from functools import partial
-from typing import IO
 from typing import Final
 
 import databento_dbn
@@ -20,7 +19,7 @@ from databento.common.constants import ALL_SYMBOLS
 from databento.common.enums import ReconnectPolicy
 from databento.common.error import BentoError
 from databento.common.publishers import Dataset
-from databento.common.types import DBNRecord
+from databento.common.types import ClientStream, DBNRecord
 from databento.common.types import ExceptionCallback
 from databento.common.types import ReconnectCallback
 from databento.common.types import RecordCallback
@@ -148,6 +147,12 @@ class SessionMetadata:
     def __bool__(self) -> bool:
         return self.data is not None
 
+    @property
+    def has_ts_out(self) -> bool:
+        if self.data is None:
+            return False
+        return self.data.ts_out
+
     def check(self, other: databento_dbn.Metadata) -> None:
         """
         Verify the Metadata is compatible with another Metadata message. This
@@ -191,7 +196,7 @@ class _SessionProtocol(DatabentoLiveProtocol):
         dataset: Dataset | str,
         dbn_queue: DBNQueue,
         user_callbacks: list[tuple[RecordCallback, ExceptionCallback | None]],
-        user_streams: list[tuple[IO[bytes], ExceptionCallback | None]],
+        user_streams: list[ClientStream],
         loop: asyncio.AbstractEventLoop,
         metadata: SessionMetadata,
         ts_out: bool = False,
@@ -210,21 +215,15 @@ class _SessionProtocol(DatabentoLiveProtocol):
         if self._metadata:
             self._metadata.check(metadata)
         else:
-            metadata_bytes = metadata.encode()
-            for stream, exc_callback in self._user_streams:
+            for stream in self._user_streams:
                 try:
-                    stream.write(metadata_bytes)
+                    stream.write(metadata.encode())
                 except Exception as exc:
-                    stream_name = getattr(stream, "name", str(stream))
                     logger.error(
-                        "error writing %d bytes to `%s` stream",
-                        len(metadata_bytes),
-                        stream_name,
+                        "error writing metadata to `%s` stream",
+                        stream.stream_name,
                         exc_info=exc,
                     )
-                    if exc_callback is not None:
-                        exc_callback(exc)
-
             self._metadata.data = metadata
         return super().received_metadata(metadata)
 
@@ -252,26 +251,20 @@ class _SessionProtocol(DatabentoLiveProtocol):
                     exc_callback(exc)
 
     def _dispatch_writes(self, record: DBNRecord) -> None:
-        if hasattr(record, "ts_out"):
-            ts_out_bytes = struct.pack("Q", record.ts_out)
-        else:
-            ts_out_bytes = b""
-
-        record_bytes = bytes(record) + ts_out_bytes
-
-        for stream, exc_callback in self._user_streams:
+        record_bytes = bytes(record)
+        ts_out_bytes = struct.pack("Q", record.ts_out) if self._metadata.has_ts_out else b""
+        for stream in self._user_streams:
             try:
                 stream.write(record_bytes)
+                stream.write(ts_out_bytes)
             except Exception as exc:
-                stream_name = getattr(stream, "name", str(stream))
                 logger.error(
-                    "error writing %d bytes to `%s` stream",
-                    len(record_bytes),
-                    stream_name,
+                    "error writing %s record (%d bytes) to `%s` stream",
+                    type(record).__name__,
+                    len(record_bytes) + len(ts_out_bytes),
+                    stream.stream_name,
                     exc_info=exc,
                 )
-                if exc_callback is not None:
-                    exc_callback(exc)
 
     def _queue_for_iteration(self, record: DBNRecord) -> None:
         self._dbn_queue.put(record)
@@ -323,7 +316,7 @@ class LiveSession:
         self._metadata = SessionMetadata()
         self._user_gateway: str | None = user_gateway
         self._user_callbacks: list[tuple[RecordCallback, ExceptionCallback | None]] = []
-        self._user_streams: list[tuple[IO[bytes], ExceptionCallback | None]] = []
+        self._user_streams: list[ClientStream] = []
         self._user_reconnect_callbacks: list[tuple[ReconnectCallback, ExceptionCallback | None]] = (
             []
         )
@@ -551,10 +544,11 @@ class LiveSession:
     def _cleanup(self) -> None:
         logger.debug("cleaning up session_id=%s", self.session_id)
         self._user_callbacks.clear()
-        for item in self._user_streams:
-            stream, _ = item
-            if not stream.closed:
+        for stream in self._user_streams:
+            if not stream.is_closed:
                 stream.flush()
+            if stream.is_managed:
+                stream.close()
 
         self._user_callbacks.clear()
         self._user_streams.clear()